requirements.txt
Django==3.0.3
djangorestframework==3.11.0
psycopg2-binary==2.8.4
python-decouple==3.1
python-dotenv==0.10.3
